Refactor to handle multiple issues etc in one message.
Refactoring said() to loop through multiple issues / commits / pull requests in
a single message, and handle each one.

I'm not entirely happy with this approach, and may refactor further, probably
normalising the "thing" type, then using a despatch table to call the
appropriate code to handle it, but this should work for now.