My initial work towards handling hooks.
This was intended to use the new hook-configuring support in GH's v3 API, but
they deprecated the auth tokens we use for logging in, so I put this on hold
ages back as it will mean overhauling the module to want username and password
instead of username and API token.

I don't much like that, TBH, but trying to get oAuth working for this  is about
as attractive as being buggered by a herde of randy wildebeest with sand for
lube whilst listening to Justin Bieber, so it's the lesser of two evils.