Ignore messages from ourselves.
Amusing problem I encountered - bot was running this module, plus the normal
Title module (which announces the title of links).

It would see a SHA, fetch the details, and announce the details - including the
link to the commit.  The Title module would see that link, and announce the
title of the link - which includes the SHA.  This module would see the SHA, and
announce the details... turtles all the way down!

(TODO: submit a patch so Title ignores stuff the bot itself says.)